+++ src/ft_num/ft_nbrlen.c
@@ -6,7 +6,7 @@
 /*   By: mboivin <mboivin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/22 13:46:03 by mboivin           #+#    #+#             */
-/*   Updated: 2020/10/06 22:35:39 by mboivin          ###   ########.fr       */
+/*   Updated: 2020/10/09 17:17:06 by mboivin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 ** as an argument (e.g., length of 100 is 3, length of 42 is 2)
 */
 
-int			ft_nbrlen(int n)
+size_t		ft_nbrlen(int n)
 {
-	int		len;
+	size_t	len;
 
 	if (n == 0)
 		return (1);
