+++ src/ft_num/ft_unbrlen_base.c
@@ -6,7 +6,7 @@
 /*   By: mboivin <mboivin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/22 13:46:03 by mboivin           #+#    #+#             */
-/*   Updated: 2020/10/06 22:35:59 by mboivin          ###   ########.fr       */
+/*   Updated: 2020/10/09 17:16:21 by mboivin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 ** as an argument using the specified base
 */
 
-int			ft_unbrlen_base(unsigned long long n, int base)
+size_t		ft_unbrlen_base(int64_t n, int base)
 {
-	int		len;
+	size_t	len;
 
 	if (n == 0)
 		return (1);
