+++ src/ft_printf/ft_handle_ptr.c
@@ -6,7 +6,7 @@
 /*   By: mboivin <mboivin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/11 13:19:02 by mboivin           #+#    #+#             */
-/*   Updated: 2020/09/30 15:11:56 by mboivin          ###   ########.fr       */
+/*   Updated: 2020/10/09 17:28:39 by mboivin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@
 ** returns: The count of printed characters
 */
 
-static int				ft_printf_ptr(char *s, t_spec *spec)
+static int		ft_printf_ptr(char *s, t_spec *spec)
 {
-	int					printed;
+	int			printed;
 
 	printed = 0;
 	if (!(spec->flag & LEFTALIGN) && spec->width > (int)ft_strlen(s))
@@ -49,14 +49,14 @@ static int				ft_printf_ptr(char *s, t_spec *spec)
 ** returns: The count of printed characters
 */
 
-int						ft_handle_ptr(t_spec *spec, va_list ap)
+int				ft_handle_ptr(t_spec *spec, va_list ap)
 {
-	int					printed;
-	unsigned long long	arg;
-	char				*buffer;
+	int			printed;
+	uint64_t	arg;
+	char		*buffer;
 
 	printed = 0;
-	arg = (unsigned long long)va_arg(ap, void *);
+	arg = (uint64_t)va_arg(ap, void *);
 	buffer = ft_utoa_base(arg, HEX_BASE);
 	if (spec->prec == 0)
 		buffer = ft_strdup(LOWHEX_PREFIX);
